Bug#38009073 Router logs a password of bootstrap autocreated user if log level is debug
When the Router is bootstrapped without the --account parameter, one
is created via SQL with a random password that is stored in the
keyring. If on top of that logger debug level is forced using
--logger.level=debug, each SQL is logged (by default to the console)

In the logs one can see:
127.0.0.1:6000 (11102 us)> CREATE USER IF NOT EXIST`S 'mysql_router1_9lgm9f5'@'%'
IDENTIFIED WITH `caching_sha2_password` BY 'random-string' // OK

The password should be filtered there and replaced by ***. This is how
this worked in the past while the mysql_native_password was used as an auth plugin.
Once that was switched to `caching_sha2_password` the filtering stopped properly
matching the statement.

This patch fixes the filter regex and adds a test that makes sure that
the replacement occures when --logger.level=debug is used.

Change-Id: I27e22a6270680d97741acb65c40d5cc0a8aa164a