Bug#38027636 Add support for cmake 4 [MySQL 5.7]
This patch is for 5.7, which is EOL, but we would still like to be
able to build it, to compare features/results in 5.7 vs newer
releases.

The patch disables features/policies which are no longer supported in
cmake 4.0. In particular we disable embedded server, which depends on
unsupported cmake policies.

Change-Id: I3b0d13ab66f81529e607e712840a2df65dca2403