BUG#37986380: REQUIRE_TABLE_PRIMARY_KEY_CHECK = GENERATE               breaks CREATE TABLE ... AS SELECT
DESCRIPTION:
===========
CREATE TABLE ... START TRANSACTION was incorrectly written
to binary log when sql_generate_invisible_primary_key was
enabled. As a result, replication of such statements failed
with error when applied on replica.

ANALYSIS:
=========
When binlogging a 'CREATE TABLE' query, store_create_info()
is called if the query needs to be rewritten before
binlogging. store_create_info() will build the CREATE TABLE
statement. For eg, a CTAS query needs to be binlogged in the
below format:
BEGIN
CREATE TABLE t1 (c1 INT) START TRANSACTION
Write_rows events on t1
COMMIT

Consider the query:
CREATE TABLE t1(c1 INT) START TRANSACTION;

With sql_generate_invisible_primary_key disabled, the query
will be binlogged as is.
With sql_generate_invisible_primary_key enabled, the CREATE
TABLE statement needs to be rewritten by including the
generated primary key information in the table metadata. In
store_create_info(), while checking if 'START TRANSACTION'
clause needs to be added, the current handling only checks
if the query is a CTAS and if the SE supports atomic DDL.
Thus, we miss appending the 'START TRANSACTION' clause for
the above query. As a result, on replica, when the applier
tries to execute the CREATE TABLE statement, an error is
reported since it is an implicit-committing statement in the
middle of a transaction that has an assigned GTID.

FIX:
====
Added a check to ensure 'START TRANSACTION' clause is
appended to the binlogged statement for 'CREATE TABLE ...
START TRANSACTION' query when
sql_generate_invisible_primary_key is enabled.

Change-Id: I05e2a35d9d8c129193f01cf9187ca39eeccf4a73